Add protection against HTTP header fingerprinting
This forces clients to send only the required headers to perform the
request and no additional headers. If additional headers are sent, the
request will be rejected with the HTTP status code 500 (Bad Request).

Of course the P-Service could still perform HTTP header fingerprinting
attacks, even if it is rejecting the request (which can be validated
externally). But the intention is to force clients to avoid sending
superfluous HTTP headers in the first place, as the requests will fail
else which would render using the service meaningless.