First version of pservice using the Bottle python web framework.
Compared to the previous version, this rewrite is more efficient
and can be configured to handle more load. The main advantage is
the usage of one of Python's WSGI compliant web frameworks, which
enables us to use any python server desired. In this case, the
gunicorn web server is preferred.

Other changes compared to the previous version:
- TLS support removed from application itself. This should be
handled by proxy such as nginx.
- Configuration is now done using a the conf.json file, rather
than using command line switches. This was necessary to allow
the service to be compatible with a multitude of python servers.
- Events can now be logged. Event logging includes the full HTTP
request along with all of its headers, a timestamp, sender IP
information, and stored values before and after the requested
operation.